Reorganize builtins tests (#35160)
