+++ test/integration/builtins/test-amp-pixel.js
