+++ test/unit/builtins/test-amp-img.js
@@ -19,10 +19,10 @@ import {
   AmpImg,
   installImg,
 } from '#builtins/amp-img/amp-img';
-import {BaseElement} from '../../src/base-element';
+import {BaseElement} from '../../../src/base-element';
 import {Layout, LayoutPriority} from '#core/dom/layout';
 import {Services} from '#service';
-import {createCustomEvent} from '../../src/event-helper';
+import {createCustomEvent} from '../../../src/event-helper';
 import {createIframePromise} from '#testing/iframe';
 
 describes.sandboxed('amp-img', {}, (env) => {
