+++ test/unit/builtins/test-amp-pixel.js
